+++ src/executables/mod.rs
@@ -4,10 +4,12 @@ mod command_info;
 mod executable;
 mod executable_call;
 mod executable_name;
+mod load_or_install;
 mod run_method;
 
 pub use command_info::CommandInfo;
 pub use executable::Executable;
 pub use executable_call::{ExecutableArgs, ExecutableCall, ExecutableCallDefinition};
 pub use executable_name::{ExecutableNamePlatform, ExecutableNameUnix};
+pub use load_or_install::{load_or_install_app, load_or_install_apps};
 pub use run_method::RunMethod;
