+++ src/lib.rs
@@ -59,10 +59,10 @@ mod strings;
 mod subshell;
 mod yard;
 use crate::applications::{AppDefinition, Apps, carrier};
-use crate::commands::{load_or_install_app, load_or_install_apps};
-use crate::configuration::{AppVersions, RequestedVersions};
+use crate::configuration::RequestedVersions;
 use crate::context::RuntimeContext;
 pub use crate::executables::CommandInfo;
+use crate::executables::{load_or_install_app, load_or_install_apps};
 use crate::yard::Yard;
 use cli::Cli;
 pub use configuration::Version;
